OneDrive/Documents/PalayProtectorV3/requirements.txt
streamlit
requests
pillow
supabase
python-dotenv
numpy
supabase
matplotlib